requirements.txt
ujson==5.8.0
web3==6.8.0
lru-dict==1.2.0
matplotlib==3.7.2
python-dotenv==1.0.0
psycopg==3.1.10
psycopg-binary==3.1.10
bloom-filter2==2.0.0
backoff==2.2.1
tabulate==0.9.0
scipy==1.14.0
